# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RAW.Ninja is an application to store RAW image files.",
        "title": "RAW.Ninja API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/account/login": {
            "post": {
                "description": "Logs in the user, sets up the JWT authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User login endpoint",
                "parameters": [
                    {
                        "description": "Credentials provided for the login",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
...
        "/roles/": {
            "get": {
                "description": "Lists all roles of the application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Role list endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/roles/:id": {
            "put": {
                "description": "Updates the settings of a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Role update endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the role information to patch",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The new version of the role to use for update",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/role.ProfileRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "description": "Lists the users of the application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List users endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/users/:id": {
            "put": {
                "description": "Updates the target user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User update endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user information to patch",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The new version of the user information to use for update",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Profile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the target user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User update endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user information to patch",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/users/:id/enabled": {
            "put": {
                "description": "Updates the target user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User enable/disable endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user information to patch",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Whether the user is enabled to log in and upload photos",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SetEnabled"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "description": "Gets the current logged in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get user profile endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Profile"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.StatusMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "role.ProfileRole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quota": {
                    "type": "integer"
                }
            }
        },
...
        "user.Credentials": {
            "type": "object",
            "required": [
                "captcha_token",
                "email",
                "password"
            ],
            "properties": {
                "captcha_token": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.Profile": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/role.ProfileRole"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "user.SetEnabled": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    },
...
}